ip2gid: Move log related code to a separate file
Move log related code to log.c and log.h so that log APIs can be used
in further patches.
